Ensure notification entities brought in by query are tracked as Unchanged
Issue #4020

The issue is that when query sets the navigation property this is detected by the change tracker and the new entity is marked as Added. There doesn't seem to be a clean way to prevent this, although it possibly could be done by looking at the stack or by setting some kind of thread-local flag to tell the change tracker that query is making the changes. Instead, when query is done and tells EF to start tracking the entity we now explicitly set the state to Unchanged even if the entity is already being tracked.